Revert "Scaffold Wallets (#8)"
This reverts commit b49ec1acd5d38a39b51a0c5c284fcf59bf1daf5b.
