+++ Tests/SolanaRPCTests/SolanaRPCTests.swift
@@ -2,3 +2,8 @@ import Foundation
 import Testing
 
 @testable import SolanaRPC
+
+@Test func myTest() async {
+    let client = SolanaRPCClient(endpoint: .devnet)
+    print(try! await client.getVersion())
+}
